chore!: Inline guide-key and undo-tree modules
`undo-tree` and `guide-key` are now configured by the personal
configuration and do not exist as module anymore.
The line between my personal configuration and my modules are narrow
and unclear, but currently I think that moving stuff that I don't want
to disable to my config makes sense.

BREAKING CHANGE: configuration of guide-key and undo-tree is moved to
README.org instead of individual modules.